+++ CHANGELOG.md
@@ -6,6 +6,9 @@
   [(#140)](https://github.com/PennyLaneAI/PennyLane-IonQ/pull/140)
 
 ### Improvements 🛠
+ 
+* The Pennylane Evolution gate is now decomposed to IonQ pauliexp gates before being sent to IonQ API.
+  [(#134)](https://github.com/PennyLaneAI/PennyLane-IonQ/pull/134)
 
 ### Breaking changes 💔
 
@@ -22,6 +25,7 @@
 
 This release contains contributions from (in alphabetical order):
 
+Radu Marginean,
 Andrija Paurevic.
 Radu Marginean
 Spencer Churchill